requirements.txt
aiofiles==0.4.0
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
cffi==1.14.0
chardet==3.0.4
future==0.18.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
idna==2.10
importlib-metadata==1.7.0
jsonschema==3.2.0
Logbook==1.5.3
matrix-nio==0.14.1
multidict==4.7.6
pycparser==2.20
pycryptodome==3.9.8
pyrsistent==0.16.0
six==1.15.0
unpaddedbase64==1.1.0
yarl==1.4.2
zipp==3.1.0
